iOS application made from scratch in Swift UI
This iOS app connects to a Firebase database, reads live data from the database, and displays the information on the app. It also features a log-in screen with authentication safety metrics, allowing users to create an account with an email and password. Multiple different users can have their information stored in the database, which can be accessed forever. Page that reads data from firebase has a refresh button, when pressed, reads the most current up to date information available.